Add tests and fix recipe attributes error.
Add Chefspec tests.

[Chefspec](https://github.com/sethvargo/chefspec) is an rspec extension
for chef recipe unit testing.

It allows you to develop the cookbook in a TDD manner, and without having to
run `vagrant up` to see recipe compilation issues.

Add TestKitchen Serverspec tests

[TestKitchen](http://kitchen.ci/) Allows you to write integration tests
so that you can verify your install on a number of different platforms.

[Serverspec](http://serverspec.org/) provides integration test resources
in an rspec syntax.